Clean up ANSI warnings
git-svn-id: https://svn.so-much-stuff.com/svn/trunk/cvs/trunk@1408 93276802-5317-4dee-a4ef-7ae40ebda4b9

+++ games.d/nethack.d/rip.c
@@ -46,7 +46,7 @@ outrip(){
 		index(vowels, *killer) ? " an" : " a");
 	center(8, buf);
 	(void) strcpy(buf, killer);
-	if(strlen(buf) > 16) {
+	if(strlen(buf) > (unsigned)16) {
 	    register int i,i0,i1;
 		i0 = i1 = 0;
 		for(i = 0; i <= 16; i++)
@@ -80,6 +80,6 @@ outrip(){
 center(line, text) int line; char *text; {
 register char *ip,*op;
 	ip = text;
-	op = &rip[line][28 - ((strlen(text)+1)/2)];
+	op = &rip[line][28 - ((strlen(text)+1)/(unsigned)2)];
 	while(*ip) *op++ = *ip++;
 }
